LastMeetingThreshold extracted to a component
