+++ front-end/src/App.js
@@ -82,7 +82,7 @@ function App() {
         setBackendFetchError(error);
       });
 
-    api({endpoint: `/config`})
+    api({endpoint: `/config`, token: idToken})
     .then(response => response.json())
     .then(body => setLastMeetingThresholdsInWeeks(body.lastMeetingThresholdsInWeeks));
   }, [defaultSort, shouldRender]);
@@ -116,6 +116,7 @@ function App() {
                   craftspeople={craftspeople}
                   rerender={rerender}
                   idToken={idToken}
+                  lastMeetingThresholdDefaultValue={lastMeetingThresholdsInWeeks}
                 />
               </Col>
             </Row>
